refactor: structuring the useGetData as a hook to be called on the giftList page
